conda-lineage-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiobotocore=2.22.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aioitertools=0.12.0=pypi_0
aiosignal=1.3.2=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
boto3=1.37.3=pypi_0
botocore=1.37.3=pypi_0
ca-certificates=2025.2.25=hecd8cb5_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
frozenlist=1.6.0=pypi_0
fsspec=2024.12.0=pypi_0
idna=3.10=pypi_0
jmespath=1.0.1=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_1
multidict=6.4.4=pypi_0
ncurses=6.4=hcec6c5f_0
numpy=2.0.2=pypi_0
openlineage-python=1.26.0=pypi_0
openssl=3.0.16=h184c1cd_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
pip=25.1=pyhc872135_2
propcache=0.3.1=pypi_0
python=3.9.21=hce00570_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=hca72f7f_0
requests=2.32.3=pypi_0
s3fs=2024.12.0=pypi_0
s3transfer=0.11.3=pypi_0
setuptools=78.1.1=py39hecd8cb5_0
six=1.17.0=pypi_0
sqlite=3.45.3=h6c40b1e_0
tk=8.6.14=h4d00af3_0
typing-extensions=4.13.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=1.26.20=pypi_0
wheel=0.45.1=py39hecd8cb5_0
wrapt=1.17.2=pypi_0
xz=5.6.4=h46256e1_1
yarl=1.20.0=pypi_0
zlib=1.2.13=h4b97444_1
